Exericse #5 fix

+++ pointers_arrays_strings/5-rev_string.c
@@ -13,7 +13,7 @@ void rev_string(char *s)
 	char c;
 	int i;
 
-	for (i = 0; i < strlen(s) / 2; i++)
+	for (i = 0; i < ((int) strlen(s)) / 2; i++)
 	{
 		c = s[strlen(s) - i - 1];
 		s[strlen(s) - i - 1] = s[i];
